Add files via upload
I still need to get the donut at the end and the find line protocol but I successfully completed the course on 8/8 trial runs!!! Should be able to finish this up by Wednesday then proceed to clean it up and compartmentalize it more. Note: watch the implicit type conversions, I was losing data because one of my formulas had (1513/1000) instead of (1513.0/1000) we needed the implicit float conversion to allow the abs() function to maintain the data when the minimum subtraction takes the value below zero (because the array has to be unsigned ints, I already tried to make it an array of signed ints, no go due to the imported sensor functions requiring unsigned int parameters). Anywho Im gonna call it a night.